Move print_fdt method to device tree crate.
This is generic method that can be used by any mainboard so moving it to
more common location (so that other mainboards can use it as well).
